refactor(swc/ecma_ast): align exports to ecma_ast
