+++ crates/swc_common/src/private/mod.rs
@@ -1,3 +1,3 @@
 //! This module is private module and can be changed without notice.
 
-pub use ::serde::__private as serde;
+pub use serde::__private as serde;
