+++ crates/swc_ecma_ast/src/list.rs
@@ -1,8 +1,7 @@
 #![allow(unused)]
 #![allow(non_upper_case_globals)]
 use bitflags::bitflags;
-
-use crate::add_bitflags;
+use swc_common::add_bitflags;
 
 bitflags! {
     pub struct ListFormat: u32 {
